Separate types from utils
