+++ src/algorithms/types/TimeSeries.types.ts
@@ -1,0 +1,6 @@
+interface TimePoint {
+  t: Date
+  y: number
+}
+
+export type TimeSeries = TimePoint[]
