+++ src/algorithms/utils/TimeSeries.ts
@@ -2,13 +2,7 @@ import * as d3 from 'd3'
 
 import { DateRange } from '../types/Param.types'
 import { interpolateTimeSeries } from '../run'
-
-interface TimePoint {
-  t: Date
-  y: number
-}
-
-export type TimeSeries = TimePoint[]
+import { TimeSeries } from '../types/TimeSeries.types'
 
 export function uniformDatesBetween(min: number, max: number, n: number): Date[] {
   const d     = (max - min) / (n - 1)
